Deleted all folders from DB and sql.py
Fixed a lot of errors.

Made connection a bit more secure and less errors
when trying to run for the first time.

Finished all SQLite needed commands (As far as I know)
Added a exit command for TUI.

The program now creates all necessary folders for
the program to correctly work.

When checking the date it now will remove all songs
with a countdown value of 0 and will reduce by 1
all remaining songs' countdown.

Removed unnecessary comments on youtube.py.